+++ app/panelists/reports/gender_stats.py
@@ -117,9 +117,9 @@ def retrieve_stats_by_year_gender(
                     all_stats[year][gender] = {
                         "minimum": Decimal(numpy.amin(scores)),
                         "maximum": Decimal(numpy.amax(scores)),
-                        "mean": Decimal(numpy.mean(scores)),
+                        "mean": round(Decimal(numpy.mean(scores)), 5),
                         "median": Decimal(numpy.median(scores)),
-                        "standard_deviation": Decimal(numpy.std(scores)),
+                        "standard_deviation": round(Decimal(numpy.std(scores)), 5),
                         "count": Decimal(len(scores)),
                         "total": Decimal(numpy.sum(scores)),
                     }
@@ -127,9 +127,9 @@ def retrieve_stats_by_year_gender(
                     all_stats[year][gender] = {
                         "minimum": int(numpy.amin(scores)),
                         "maximum": int(numpy.amax(scores)),
-                        "mean": round(numpy.mean(scores), 4),
+                        "mean": round(numpy.mean(scores), 5),
                         "median": int(numpy.median(scores)),
-                        "standard_deviation": round(numpy.std(scores), 4),
+                        "standard_deviation": round(numpy.std(scores), 5),
                         "count": len(scores),
                         "total": int(numpy.sum(scores)),
                     }